refactor-realtime-positions * Move useLinesWithRealtimePositions.ts to it's own directory * Move useRealtimeVehicleReducer.ts to use-realtime-vehicle-data directory * Remove unused fields from VehicleFragment
