+++ src/logic/use-lines-with-realtime-positions/useLinesWithRealtimePositions.vehicles.graphql
